+++ src/main/java/org/md/api/auth/repository/ICredentialRepository.java
@@ -4,4 +4,6 @@ import org.md.api.auth.model.exception.InvalidCredentialsException;
 
 public interface ICredentialRepository {
     public void secureCredentialsAreValid(String username, String password) throws InvalidCredentialsException;
+
+    public boolean repositoryIsLive() throws Exception;
 }